move ImportMode and AddStrategyParameters to DTO
