+++ app/Console/Commands/Sync.php
@@ -3,8 +3,8 @@
 namespace App\Console\Commands;
 
 use App\Actions\Import\Exec;
-use App\Actions\Photo\Strategies\ImportMode;
 use App\Contracts\Exceptions\ExternalLycheeException;
+use App\DTO\ImportMode;
 use App\Exceptions\ConfigurationKeyMissingException;
 use App\Exceptions\UnexpectedException;
 use App\Models\Album;
